pyproject.toml
[project]
name = "kosmos_meteor"
version = "0.1.0"
description = "A Python model for Probabilistic Asteroid Impact Risk assessment."
authors = [{ name = "Your Name", email = "your.email@example.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Astronomy",
]

# Core dependencies required for the simulation
dependencies = [
    "numpy",
    "scipy",
    "pandas",
    "xarray",
    "h5py", # For efficient storage of large datasets
    "numba", # For accelerating critical numerical functions
    "tqdm", # For progress bars
    "streamlit>=1.50.0",
    "plotly>=6.3.0",
]

[project.optional-dependencies]
# Dependencies for visualization and webapp
viz = [
    "matplotlib",
    "seaborn",
    "plotly",
    "streamlit",
    "folium",
    "streamlit-folium",
    "simplekml",     # For generating KML files
    "pyyaml",
]

# Dependencies for building documentation
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
    "pymdown-extensions",
    "mkdocs-mermaid2-plugin",
]

# A single group to install everything
all = [
    "kosmos_meteor[viz]",
    "kosmos_meteor[docs]",
]

[project.urls]
Homepage = "https://github.com/your-username/PyPAIR"
Repository = "https://github.com/your-username/PyPAIR"

[tool.uv.sources]
# You can add private package indexes here if needed
# my-index = { url = "https://my-index.com/simple" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
